server environment: centralize various env settings
currently, we invoke `process.env` all over the place, and it's unclear
which environment settings are active where. this creates a centralized
repository for env vars/env-specific settings. we immediately put it to
use with a new `DATA_DIR` setting, which would replace the existing
combo of `UPLOAD_DIRECTORY` and `TREASURY_DIRECTORY`