overhaul setting of `REPORTING_TEMPLATE`
we handled this in several places, and now it's basically handled in
two. the two places are in `environment.js`, where we configure the
default empty template name, and in the routes for `reporting_period`,
where we both upload and download the template for the given reporting
period.

we stop looking at the `reporting_template` column of the application
settings -- this was never used anyway.

we stop using the `REPORTING_TEMPLATE` env var -- this doesn't really
change, if we want to change it we should change the code.

we kill the `/template` routes, and use the reporting period API for
template management.