persist uploads separately from validating them
lots of changes in this commit -- sorry, this is a bit of a hairball.

there's a migration to drop some redundant columns on the uploads and
documents tables, including the extra user id on documents, the project
id on uploads (which makes no sense now, since an upload can have
multiple projects). we drop the extra email field, since you can just
get that from the associated user_id.

we change how uploads are persisted to the filesystem -- they now use
their upload id as the base of their filename -- which allows us to drop
the uniqueness constraint on the upload filenames table.

we persist the upload exactly as given, without even trying to drop rows
from the AOA structure -- we should do that on read, so we can change
how we process things in the future without requiring re-parsing the
data.

we drop a bunch of superfluous code. there is still a LOT of dead code
just generally around, and also resulting from this change, but tracking
it down will be a follow-up.