Unify validation on the `Uploads` view
we implement the validation of a new upload client-side, by bouncing you
to the new upload page and triggering a validation flow.

we can kill the existing validations UIs -- one on the NewUpload page
and one on it's own dedicated validation screen. you will now only get
an error on the NewUpload page if the upload actually failed, and
otherwise all validation errors are displayed on the upload page.

we do some client-side work to make all of this work. in particular, for
uploads, we move state management, and async fetches, out of the store
and into the component. this keeps things simpler and more unified and
will probably minimize bugs, but it's a break from the current
all-things-go-in-the-store approach.