validate-upload: begin validating agency code
we require that it be set in the file, on the cover sheet.

while we're here, we delete the `parse-filename` service, which is now
dead code. this is where the agency validation previously lived.

we refactor the agency db code into it's own file for clarity and
consistency. we fix the front-end logic around displaying validation
errors