ui: updated bootstrap for layout and navigation
we switch to the latest 4.x release of bootstrap, sourced from a
different CDN. we update the nav layout to be consistent with the
bootstrap documentation. we wrap the entire app in a flexible container
with the correct width, which eliminates custom width settings on
specific pages. we delete all custom styles on sub-views.