validate: handle transaction commit/rollback
we want to be free to make modifications that later get rolled back, for
example adding subrecipients. so, we always want to run in a
transaction. however, we also support a transaction being passed in from
the parent (e.g., for the revalidate service) in which case we allow the
parent to handle all transaction logic